Initial commit
This previously was contained in the cbmi/harvest repo under the
template/ directory. This has been moved to separate repository and
improved to not rely on Django's startproject command which enables it
to be used directly. In addition, this makes it easier to generate
diffs for upgrading existing projects.

The file named .harvestrc has been introduced for keeping track of the
version of Harvest this project represents. This is used when upgrading
to new versions of Harvest.